flask_REST_React/api/requirements.txt
flask==1.0.2
flask_script==2.0.6
flask_migrate==2.2.1
flask_sqlalchemy==2.3.2

flask_restful==0.3.6
marshmallow==3.0.0b18
flask_marshmallow==0.9.0
marshmallow-sqlalchemy==0.14.1
pytest==3.8.2

flask_cors==3.0.6